d/postinst: Change the owner of some directories and their subfiles
These directories are "/var/lib/chrony" and "/var/log/chrony". For the
latter, it is necessary to change its owner to allow chronyd to write in
the log files, which isn’t possible if we let the owner of those files
to root as chronyd is now running as "_chrony".
Note that the owner will not be changed for users having set up a custom
the "user" directive in their "chrony.conf" file. Thus, they will keep their
working configuration.