d/postrm: re-establish systemd-timesyncd on removal
This is the very tail end of the discussion we had in Debian bug 895821.
With all things resolved back then the only issue that really came up
in the interworking between systemd-timesyncd and chrony was that
install+remove of chrony would leave no time syncing up at all.

This is easily fixed by re-starting it (if enabled) in our postrm.

The change was initially using systemctl but that ignores policy-rc.d,
which is rare but exists to control services, so better use
deb-systemd-invoke. It was discussed and verified by Vincent (thanks)
that deb-systemd-invoke will check on its own if the service is masked
or disabled, so this doesn't need an extra check.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>