d/t/{dynamically-add-source,ntp-server-and-nts-auth}: reduce default ubuntu config, to make space for testcase config
In Ubuntu our default config already includes 8 servers which is all
that chrony ever needs. The autopkgtests add more in included subdirectories
which in the Ubuntu case isn't picked up.

Removing 4 of the 8 of these servers at the beginning of the autopkgtests
followed by a daemon restart makes the rest of the test code work fine.
For Debian that should be a no-op as the initial check never equals to
true, but for Ubuntu it helps a lot making the tests pass with the new
version of chrony.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>